add button reading functionality
