+++ Common/adc.c
@@ -1,0 +1,15 @@
+#include "adc.h"
+
+void adc_init() {
+    // Reference voltage: AVcc, ADC prescaler: 128
+    ADMUX = (1 << REFS0);
+    ADCSRA = (1 << ADEN) | (1 << ADPS2) | (1 << ADPS1) | (1 << ADPS0);
+}
+
+uint16_t adc_read(uint8_t channel) {
+    // Select ADC channel with safety mask
+    ADMUX = (ADMUX & 0xF8) | (channel & 0x07);
+    ADCSRA |= (1 << ADSC); // Start conversion
+    while (ADCSRA & (1 << ADSC)); // Wait for conversion to complete
+    return ADC;
+}
