+++ Common/adc.h
@@ -1,0 +1,9 @@
+#ifndef ADC_H
+#define ADC_H
+
+#include <avr/io.h>
+
+void adc_init();
+uint16_t adc_read(uint8_t channel);
+
+#endif // ADC_H
