+++ I2CMenueBasedSystem/project.c
@@ -3,21 +3,26 @@
 #include "../Common/i2c.h"
 #include "../Common/lcd.h"
 #include "../Common/uart.h"
+#include "../Common/adc.h"
 
 #define JOYSTICK_X 0 // ADC channel for X-axis
 #define JOYSTICK_Y 1 // ADC channel for Y-axis
+#define JOYSTICK_BUTTON_PIN PD2
 #define THRESHOLD 100
 
-void adc_init();
-uint16_t adc_read(uint8_t channel);
+void button_init();
+uint8_t button_read();
 
 int main(void) {
     // Initialize UART
-    adc_init(); // Initialize ADC
+    uart_init(103); // 9600 baud rate for 16MHz clock
+    uart_println("Button test");
 
-    // Initialize I2C and LCD
+    // Initialize ADC, I2C, and LCD
+    adc_init();
     i2c_init();
     lcd_init();
+    button_init();
 
     const uint16_t threshold = 200;
     const uint16_t center = 512;
@@ -35,22 +40,22 @@ int main(void) {
             lcd_print("second item");
         }
 
+        // Read joystick button state
+        if (button_read() == 0) { // Button pressed
+            uart_println("Button pressed");
+        }
+
         _delay_ms(200); // Delay for readability
     }
 
     return 0;
 }
 
-void adc_init() {
-    // Reference voltage: AVcc, ADC prescaler: 128
-    ADMUX = (1 << REFS0);
-    ADCSRA = (1 << ADEN) | (1 << ADPS2) | (1 << ADPS1) | (1 << ADPS0);
+void button_init() {
+    DDRD &= ~(1 << JOYSTICK_BUTTON_PIN); // Set button pin as input
+    PORTD |= (1 << JOYSTICK_BUTTON_PIN); // Enable pull-up resistor
 }
 
-uint16_t adc_read(uint8_t channel) {
-    // Select ADC channel with safety mask
-    ADMUX = (ADMUX & 0xF8) | (channel & 0x07);
-    ADCSRA |= (1 << ADSC); // Start conversion
-    while (ADCSRA & (1 << ADSC)); // Wait for conversion to complete
-    return ADC;
+uint8_t button_read() {
+    return PIND & (1 << JOYSTICK_BUTTON_PIN); // Read button state
 }
